+++ src/main/java/org/example/projectsagant/model/Channel.java
@@ -1,0 +1,7 @@
+package org.example.projectsagant.model;
+
+public enum Channel {
+    LOG,
+    SERVICE,
+    EMAIL
+}